refactor: 에러 핸들링 함수와 메시지 출력 함수를 utils로 통합
