--- src/common/utils/errorHandler.js
@@ -1,5 +1,0 @@
-const throwError = (message, condition = true) => {
-  if (condition) throw new Error(message);
-};
-
-export default throwError;
