--- src/common/utils/messagePrinter.js
@@ -1,5 +1,0 @@
-import { Console } from '@woowacourse/mission-utils';
-
-const printMessage = (message) => Console.print(message);
-
-export default printMessage;
